Add CI build through RISC OS Build service.
This change introduces automated building of the sources using the
RISC OS Build service - we send the code and makefile to the service
and it builds things according to the makefile, and then returns the
result to us.

We should also be able to create a tag prefixed with 'v' to create
a release automatically as well.